+++ agent/fileclient/sync.py
@@ -1,9 +1,9 @@
 import subprocess
 
 
-def sync(address: str, path: str) -> bool:
+def sync(address: str, path: str, auth: str) -> bool:
     process = subprocess.Popen(
-        ["rclone", "copy", ":http:/", "--http-url", f"http://{address}", path]
+        ["rclone", "copy", ":http:/", "--http-url", f"http://{auth}@{address}", path]
     )
     process.wait()
     return process.returncode == 0